Vagrant up (#3)
* Fixes #1

The purpose of this commit is to ensure that the `.vagrant` directory
does not inadvertantly make it into the repo and cause issues for the
team.

From an interview perspective, this commit is about showing some
knowledge of workflows activated by commit messages.

* Fixes contrast-security-oss/ops-hire-project#1

Intial `vagrant up` errors and leads to some confusion on the part of
the interviewee. Fix error so they can get to the meat of the problem.

* Fixes #11

Updated the method by which the ansible user becomes `root` on a
target machine to silence a deprecation warning.